refactor: 의존 관계 수정
