+++ src/main/java/com/example/demo/infra/board/entity/Board.java
@@ -9,11 +9,11 @@ import com.example.demo.domain.board.service.entity.BoardContent;
 import com.example.demo.domain.board.service.entity.BoardInfo;
 import com.example.demo.domain.comment.domain.entity.BoardComment;
 import com.example.demo.domain.recruitment_board.domain.entity.GenericBoard;
-import com.example.demo.domain.user.domain.User;
-import com.example.demo.domain.user.domain.UserTarget;
+import com.example.demo.domain.user.entity.UserTarget;
 import com.example.demo.global.base.domain.BaseEntity;
 import com.example.demo.infra.board.category.entity.BoardCategory;
 
+import com.example.demo.infra.user.entity.User;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
